consumer/requirements.txt
fluvio
python-socketio
